redux_toolkit: WIP refactoring
