+++ src/Container/Markdown/EditForm/NoteEditForm.tsx
@@ -5,8 +5,8 @@ import { uuid } from 'uuidv4';
 import { CloseOutlined, CheckOutlined } from '@ant-design/icons';
 import { DATE_TIME_FORMAT } from 'Constants/Common';
 import { ActionsPanel } from 'Container/Markdown/ActionsPanel';
-import MarkdownNoteContext from 'Container/Markdown/MarkdownContext';
 import { INotesSectionData } from 'Container/Markdown/Models';
+import MarkdownNoteContext from 'Container/Markdown/NotesContext';
 import styles from 'Style/MarkdownEditor/MarkdowInput.less';
 import { openNotificationWithIcon, convertStringArrayToString } from 'Utils/common';
 import { useLiteralValue } from 'Utils/hooks';
