+++ src/Store/notes/state.ts
@@ -1,31 +1,9 @@
-import moment from 'moment';
-import { uuid } from 'uuidv4';
-import { DATE_TIME_FORMAT, EState } from 'Constants/Common';
+import { EState } from 'Constants/Common';
 import { INotesSectionContext } from 'Container/Markdown/Models';
-import { EFormStep } from 'Container/Markdown/enums';
+import { defaultContext } from 'Container/Markdown/NotesContext';
 
 export const initialState: INotesSectionContext = {
-    data: {
-        currentNote: null,
-        notes: [
-            {
-                id: uuid(),
-                createDate: moment(new Date()).format(DATE_TIME_FORMAT),
-                title: 'Title',
-                text: '1ju7OW3uOhy0oYqLTiEPNg==',
-                isEncrypted: true
-            },
-            {
-                id: uuid(),
-                createDate: moment(new Date()).format(DATE_TIME_FORMAT),
-                title: 'Title',
-                text: '# Header\n>this decrypted note.',
-                isEncrypted: false
-            }
-        ],
-        step: EFormStep.LIST,
-        markdownText: ''
-    },
+    data: defaultContext,
     state: EState.NONE,
     errors: []
 };
